Rename module from 'recoverFrom_0_6_to_0_9' to 'from_0_6_to_0_9' in data recovery to fix "TypeError: 'module' object is not callable"
